Update after UX feedback
